backend/pyproject.toml
[project]
name = "python-real-estate-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.3",
    "folium>=0.20.0",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "openai>=2.31.0",
    "openpyxl>=3.1.5",
    "oracledb>=3.4.2",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "seaborn>=0.13.2",
    "streamlit>=1.56.0",
    "streamlit-folium>=0.27.1",
    "uvicorn>=0.44.0",
]
